+++ algo.py
@@ -3,16 +3,22 @@ import datetime
 import queue
 import sys
 import time
+from enum import IntEnum
 
 from loguru import logger
 
 # from rich.console import Console
 # from rich.theme import Theme
-from ib_client import IBClient, qu_ask, qu_bid, qu_ctx, qu_orderstatus, qu_pnlsingle
+from ib_client import IBClient, qu_ask, qu_bid, qu_ctx, qu_orderstatus
 from trade import Trade
 
 # pnl_theme = Theme({"gain": "green", "loss": "red"})
 # console = Console(theme=pnl_theme)
+
+
+class OrderType(IntEnum):
+    BUY = 1
+    SELL = 2
 
 
 def enter_trade(t: Trade, client: IBClient):
@@ -71,9 +77,14 @@ def enter_trade(t: Trade, client: IBClient):
 
 
 def check_order(t: Trade, client: IBClient):
+    # allow 5 attempts before cancelling order
+    x = 0
     req = f" reqid: {client.order_id} >>>"
     while True:
         try:
+            if x == 6:
+                pass
+                # send cancel order
             msg = qu_orderstatus.get(timeout=5)
             t.console.print(f" {req} Status: {msg['status']} ")
             if msg["status"] == "Filled":
@@ -82,36 +93,12 @@ def check_order(t: Trade, client: IBClient):
                 t.entry_price = msg["avgFillPrice"]
                 break
             else:
+                x = x + 1
                 continue
         except queue.Empty:
             t.console.print(f" {req} Status: Waiting fo fill order ")
             continue
-
-
-def getPnlSingle(t: Trade, client: IBClient, account: str) -> (float, float):
-    req = f" reqid: {client.order_id} >>>"
-    client.nextId()
-    # Wait for status
-    while True:
-        client.reqPnLSingle(
-            reqId=client.order_id, account=account, modelCode="", conid=t.conid
-        )
-        # wait for TWS to run callback
-        time.sleep(1)
-        try:
-            msg = qu_pnlsingle.get(timeout=5)
-            # pnl = msg["unrealizedPnL"]
-            pnl = msg["value"]
-            pnl_pct = (pnl - (t.avgFillPrice * t.position)) / pnl * 100
-            t.display()
-            t.console.print(req)
-            break
-
-        except queue.Empty:
-            t.console.print(f" {req} Status: Waiting pnl data ")
-            continue
-
-    return pnl_pct
+            x = x + 1
 
 
 def exit_trade(t: Trade, client: IBClient):
