+++ todo_webapp/todo.md
@@ -1,0 +1,2 @@
+1. add stop loss
+2. add modify order if exit trade is not executed in 3 seconds
